+++ src/main/java/com/aerotrack/AerotrackApp.java
@@ -1,10 +1,8 @@
-package com.myorg;
+package com.aerotrack;
 
 import software.amazon.awscdk.App;
 import software.amazon.awscdk.Environment;
 import software.amazon.awscdk.StackProps;
-
-import java.util.Arrays;
 
 public class AerotrackApp {
     public static void main(final String[] args) {
